SmartHealthMonitoringSystem_Group19/LiveMonitoringDashboard/requirements.txt
paho-mqtt
flask
flask-socketio
eventlet
sqlalchemy